+++ cloud-collect-web/src/types/index.d.ts
@@ -1,4 +1,12 @@
-//REGION entity
+//REGION Frontend
+
+export interface NavItem {
+  index: string
+  path: string
+  name: string
+}
+
+//REGION Backend
 
 //NOTE 这里不能直接使用Collect作为接口名，因为与组件名重复了，是否可考虑使用ICollect？