+++ cloud-collect-web/src/views/error/Error403View.vue
@@ -1,12 +1,12 @@
 <template>
   <!--话说这里的语言注入到底是怎么定位的……-->
-  <error>
+  <Error>
     <template v-slot:title>权限错误</template>
 
-    <i class="el-icon-warning"></i>
+    <ElIcon name="warning"/>
     这片领域不属于你……
-    <i class="el-icon-warning"></i>
-  </error>
+    <ElIcon name="warning"/>
+  </Error>
 </template>
 
 <script lang="ts">